ads integration library
