+++ test/integration/BUILD
@@ -17,13 +17,45 @@ load(
 
 envoy_package()
 
-envoy_cc_test(
-    name = "ads_integration_test",
-    srcs = ["ads_integration_test.cc"],
+envoy_cc_test_library(
+    name = "ads_integration_lib",
+    srcs = [
+        "ads_integration.cc",
+    ],
+    hdrs = [
+        "ads_integration.h",
+    ],
     data = [
         "//test/config/integration/certs",
     ],
     deps = [
+        ":http_integration_lib",
+        "//source/common/config:protobuf_link_hacks",
+        "//source/common/config:resources_lib",
+        "//source/common/protobuf:utility_lib",
+        "//source/extensions/transport_sockets/tls:config",
+        "//source/extensions/transport_sockets/tls:context_config_lib",
+        "//source/extensions/transport_sockets/tls:context_lib",
+        "//source/extensions/transport_sockets/tls:ssl_socket_lib",
+        "//test/common/grpc:grpc_client_integration_lib",
+        "//test/mocks/runtime:runtime_mocks",
+        "//test/mocks/server:server_mocks",
+        "//test/test_common:network_utility_lib",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/api/v2:cds_cc",
+        "@envoy_api//envoy/api/v2:discovery_cc",
+        "@envoy_api//envoy/api/v2:eds_cc",
+        "@envoy_api//envoy/api/v2:lds_cc",
+        "@envoy_api//envoy/api/v2:rds_cc",
+        "@envoy_api//envoy/service/discovery/v2:ads_cc",
+    ],
+)
+
+envoy_cc_test(
+    name = "ads_integration_test",
+    srcs = ["ads_integration_test.cc"],
+    deps = [
+        ":ads_integration_lib",
         ":http_integration_lib",
         "//source/common/config:protobuf_link_hacks",
         "//source/common/config:resources_lib",
